==> Code/NeuroKG/2_enrich_KG.Rmd <==
---
title: Enrich Knowledge Graph with scRNA-seq Data
subtitle: Ayush Noori
output:
  prettydoc::html_pretty:
    theme: cayman
    highlight: github
editor_options: 
  chunk_output_type: inline
---

```{r setup, include = FALSE}
# knitr::opts_chunk$set(eval = FALSE)
```

# Dependencies

Load requisite packages and define directories.

```{r load-packages, message=FALSE, warning=FALSE}
# data manipulation
library(data.table)
library(purrr)
library(magrittr)

# plotting library
library(ggplot2)

# relative file paths
library(here)
```

Note that directories are relative to the R project path.

```{r define-directores}
# set directories
primeKG_dir = here("Data", "PrimeKG")
neuroKG_dir = here("Data", "NeuroKG")
cell_type_dir = here("Data", "CELLxGENE", "linnarsson", "processed")
```

# Read Data

First, read in NeuroKG intermediate objects from `1_filter_KG.Rmd`.

```{r read-neurokg}
# read NeuroKG filtered for brain expression and anatomy
neuroKG_nodes = readRDS(here(neuroKG_dir, "1_filter_KG", "neuroKG_nodes.RDS"))
neuroKG_edges = readRDS(here(neuroKG_dir, "1_filter_KG", "neuroKG_edges.RDS"))

# print number of nodes and edges
message("Original NeuroKG Nodes:\t", nrow(neuroKG_nodes))
message("Original NeuroKG Edges:\t", nrow(neuroKG_edges))
```

Next, read in metadata of analyzed scRNA-seq data from Linnarsson *et al.*. Remember that:
* Supercluster markers (*i.e.*, markers that define each supercluster, or cell type) are stored in `supercluster_markers.csv`.
* Cluster markers (*i.e.*, markers that define clusters within each cell type) are stored in `<supercluster_name>_markers.csv`.
* Subcluster markers (*i.e.*, markers that define subclusters within each cluster) for cluster $n$ are stored in `cluster_n_markers.csv`.

```{r read-linnarsson}
# read for 2.4M neurons
neurons_genes = fread(here(cell_type_dir, "neurons_var_gene_metadata.csv"))
neurons_cells = fread(here(cell_type_dir, "neurons_obs_cell_metadata.csv"))

# read for 885K non neurons
non_neurons_genes = fread(here(cell_type_dir, "non_neurons_var_gene_metadata.csv"))
non_neurons_cells = fread(here(cell_type_dir, "non_neurons_obs_cell_metadata.csv"))

# get list of neuronal superclusters and clusters
neuron_marker_files = list.files(here(cell_type_dir, "neuron_markers"), pattern = "_markers\\.csv$")
neuron_subcluster_marker_files = neuron_marker_files[grepl("^cluster.*", neuron_marker_files)]
neuron_cluster_marker_files = neuron_marker_files[!grepl("(^cluster.*)|(^supercluster.*)", neuron_marker_files)]

# get list of non-neuronal superclusters and clusters
non_neuron_marker_files = list.files(here(cell_type_dir, "non_neuron_markers"), pattern = "_markers\\.csv$")
non_neuron_subcluster_marker_files = non_neuron_marker_files[grepl("^cluster.*", non_neuron_marker_files)]
non_neuron_cluster_marker_files = non_neuron_marker_files[!grepl("(^cluster.*)|(^supercluster.*)", non_neuron_marker_files)]

# functions to get names of files
get_supercluster_names = function(supercluster_files) { gsub("_markers\\.csv$", "", supercluster_files) }
get_cluster_names = function(cluster_files) { map_chr(cluster_files, ~paste("cluster", strsplit(.x, "_")[[1]][2])) }

# set names
names(neuron_cluster_marker_files) = get_supercluster_names(neuron_cluster_marker_files)
names(neuron_subcluster_marker_files) = get_cluster_names(neuron_subcluster_marker_files)
names(non_neuron_cluster_marker_files) = get_supercluster_names(non_neuron_cluster_marker_files)
names(non_neuron_subcluster_marker_files) = get_cluster_names(non_neuron_subcluster_marker_files)
```

For each cell type, calculate the proportion of brain regions that the cell type is present in (by number of cells). Then, construct edges between that cell type and the top 5 brain regions present in that brain region.

```{r calculate-proportions}
# subset table for neurons
neuron_cluster_region = neurons_cells[, .(ROIGroupFine, roi, dissection, tissue_ontology_term_id, supercluster_term, cell_type_ontology_term_id, cluster_id, subcluster_id)] %>%
  .[, Dataset := "Neuron"]

# subset table for non-neurons
non_neuron_cluster_region = non_neurons_cells[, .(ROIGroupFine, roi, dissection, tissue_ontology_term_id, supercluster_term, cell_type_ontology_term_id, cluster_id, subcluster_id)] %>%
  .[, Dataset := "Non-Neuron"]

# combine two data frames
cluster_region = rbind(neuron_cluster_region, non_neuron_cluster_region)

# make mapping from ROI to human readable names manually
roi_label_mapping = unique(cluster_region[, .(roi, dissection)])

# function to clean dissection name
clean_dissection = function(dissection) {
  split = strsplit(dissection, " - ")[[1]]
  label = split[length(split) - 1]
  label = tolower(label)
  if(length(label) == 1) return(label) else return(NA)
}
roi_label_mapping[, label := map(dissection, clean_dissection)]

# don't overwrite!
# fwrite(roi_label_mapping, here(neuroKG_dir, "2_enrich_KG", "ROI_to_label_mapping.csv"))
roi_label_mapping = fread(here(neuroKG_dir, "2_enrich_KG", "ROI_to_label_mapping.csv"))

# to check, see roi_label_mapping[roi %in% roi_label_mapping[duplicated(roi), roi]
roi_label_mapping = unique(roi_label_mapping[, .(roi, label)])

# dissection column is messy, so drop
# check, for example, as.data.table(tstrsplit(neuron_cluster_region$dissection, " - ", fixed = TRUE))
cluster_region[, dissection := NULL]

# get brain region to ROI mapping
region_to_roi = cluster_region %>%
  .[, .(ROIGroupFine, roi)] %>%
  unique()

# get cell type (cluster) to dissection (roi) mapping
cluster_to_roi = cluster_region[, .N, by = c("Dataset", "cluster_id", "roi")] %>%
  .[order(cluster_id, -N)] %>%
  .[, Fraction := N/sum(N)*100, by = "cluster_id"]

# check that clusters are distinct across neurons vs. non-neurons
error_clusters = intersect(cluster_to_roi[Dataset == "Neuron", cluster_id], cluster_to_roi[Dataset == "Non-Neuron", cluster_id])
# check length(error_clusters) == 0

# create histogram
cluster_to_roi_hist = ggplot(cluster_to_roi[Fraction < 15], aes(x = Fraction, fill = Dataset)) +
  geom_histogram(col="black", alpha = .5, binwidth = 1) +
  scale_fill_manual(values = c("#20A39E", "#FF715B")) +
  geom_vline(xintercept = 4.5, linetype = "dashed", color = "red") +
  facet_wrap(Dataset ~ ., ncol = 1, scales = "free_y") +
  labs(x = "Fraction", y = "Count") +
  scale_x_continuous(expand = expansion(c(0.01, 0.01)), n.breaks = 15) +
  scale_y_continuous(expand = expansion(c(0, 0.05))) +
  theme_bw() +
   theme(axis.title = element_text(face = "bold"),
         strip.text = element_text(face = "bold"),
         legend.position = "none")

# write to file
ggsave(here(neuroKG_dir, "2_enrich_KG", "cluster_to_ROI_histogram.pdf"), cluster_to_roi_hist, width = 8, height = 6)
```

# Construct Edges

We construct the following edges:
* Brain region (*i.e.*, `ROIGroupFine`) to brain structure (*i.e.*, `roi`).
* Brain structure to cell cluster.
* Cell type (*i.e.*, `supercluster`) to cell cluster.
* Cell cluster to cell subcluster.
* Cell type to marker gene (top 200).
* Cell cluster to marker gene (top 50).
* Cell subcluster to marker gene (top 10).

## Brain Region Structural Edges

First, construct brain region to brain structure (*i.e.*, `roi`) edges.

```{r region-structure}
# get new column name order
new_colorder = colnames(neuroKG_edges) %>%
  .[!(. %in% c("x_index", "y_index"))]

# construct region to structure edges
# note that x_index and y_index are added later
region_to_structure_edges = copy(region_to_roi) %>%
  merge(., roi_label_mapping, by = "roi", sort = F) %>%
  setnames(c("ROIGroupFine", "roi", "label"), c("x_id", "y_id", "y_name")) %>%
  .[, relation := "region_structure"] %>%
  .[, display_relation := "parent-child"] %>%
  .[, x_name := x_id] %>%
  .[, x_type := "brain_region"] %>%
  .[, x_source := "Siletti et al."] %>%
  .[, y_type := "brain_structure"] %>%
  .[, y_source := "Siletti et al."] %>%
  setcolorder(new_colorder)

# print number of new edges created
message("Unique Brain Regions: ", length(unique(cluster_region$ROIGroupFine)))
message("Unique Brain Structures: ", length(unique(cluster_region$roi)))
message("Brain Region --> Structure Edges: ", nrow(region_to_structure_edges))
```

Next, using the histogram from the previous step to determine the selection threshold, construct brain structure (*i.e.*, `roi`) to cell cluster edges.

```{r structure-cluster}
# construct edges
structure_to_cell_cluster_edges = cluster_to_roi[Fraction > 5, .(cluster_id, roi)] %>%
  merge(., roi_label_mapping, by = "roi", sort = F) %>%
  setnames(c("cluster_id", "roi", "label"), c("y_id", "x_id", "x_name")) %>%
  .[, relation := "structure_cell_type"] %>%
  .[, display_relation := "cell type present"] %>%
  .[, x_type := "brain_structure"] %>%
  .[, x_source := "Siletti et al."] %>%
  .[, y_type := "cell_cluster"] %>%
  .[, y_name := paste("cluster", y_id)] %>%
  .[, y_source := "Siletti et al."] %>%
  setcolorder(new_colorder)

# print number of new edges created
message("Unique Brain Structures: ", length(unique(cluster_region$roi)))
message("Unique Cell Clusters: ", length(unique(cluster_region$cluster_id)))
message("Structure --> Cluster Edges: ", nrow(structure_to_cell_cluster_edges))
```

## Cell Clustering Structural Edges

Using the `cluster_region` table created previously, construct cell type (*i.e.*, `supercluster`) to cell cluster edges.

```{r type-cluster}
cell_type_to_cell_cluster_edges = cluster_region[, .(supercluster_term, cluster_id)] %>%
  unique() %>%
  setnames(c("supercluster_term", "cluster_id"), c("x_id", "y_id")) %>%
  .[, relation := "cell_type_cell_cluster"] %>%
  .[, display_relation := "parent-child"] %>%
  .[, x_type := "cell_type"] %>%
  .[, x_name := x_id] %>%
  .[, x_source := "Siletti et al."] %>%
  .[, y_type := "cell_cluster"] %>%
  .[, y_name := paste("cluster", y_id)] %>%
  .[, y_source := "Siletti et al."] %>%
  setcolorder(new_colorder)

# print number of new edges created
message("Unique Clusters: ", length(unique(cluster_region$cluster_id)))
message("Cell Type --> Cluster Edges: ", nrow(cell_type_to_cell_cluster_edges))
```

Using the `cluster_region` table created previously, construct cell cluster to cell subcluster edges.

```{r cluster-subcluster}
cell_cluster_to_cell_subcluster_edges = cluster_region[, .(cluster_id, subcluster_id)] %>%
  unique() %>%
  setnames(c("cluster_id", "subcluster_id"), c("x_id", "y_id")) %>%
  .[, relation := "cell_cluster_cell_subcluster"] %>%
  .[, display_relation := "parent-child"] %>%
  .[, x_type := "cell_cluster"] %>%
  .[, x_name := paste("cluster", x_id)] %>%
  .[, x_source := "Siletti et al."] %>%
  .[, y_type := "cell_subcluster"] %>%
  .[, y_name := paste("subcluster", y_id)] %>%
  .[, y_source := "Siletti et al."] %>%
  setcolorder(new_colorder)

# print number of new edges created
message("Unique Subclusters: ", length(unique(cluster_region$subcluster_id)))
message("Cell Cluster --> Subcluster Edges: ", nrow(cell_cluster_to_cell_subcluster_edges))
```

## Cell Clustering Marker Gene Edges

We construct edges between genes and either cell type (*i.e.*, supercluster), cluster, or subcluster using the results of the marker gene analysis.

First, construct cell type (*i.e.*, supercluster) to marker gene edges.

```{r type-gene}
# get top n genes from each cluster
get_top_marker_genes = function(fread_result, n_genes = 200) {
  fread_result %>%
    .[1:n_genes, .SD, .SDcols = grep("_names", colnames(.), fixed = T, value = T)] %>%
    melt(measure.vars = colnames(.), variable.name = "x_id", value.name = "y_name") %>%
    return()
}

# dataset can be "neuron" or "non_neuron"
get_cell_type_to_marker_gene_edges = function(dataset = "neuron") {
  
  # file name to read supercluster markers
  supercluster_marker_file = here(cell_type_dir, paste0(dataset, "_markers"), "supercluster_markers.csv")
  
  # get top 200 genes
  supercluster_markers = fread(supercluster_marker_file) %>%
    get_top_marker_genes(n_genes = 400) %>%
    .[, relation := "cell_type_marker_gene"] %>%
    .[, display_relation := "expression present"] %>%
    .[, x_id := gsub("_names", "", x_id, fixed = T)] %>%
    .[, x_type := "cell_type"] %>%
    .[, x_name := x_id] %>%
    .[, x_source := "Siletti et al."] %>%
    merge(neuroKG_nodes[node_type == "gene/protein"], by.x = "y_name", by.y = "node_name", all.x = F, all.y = F, sort = F) %>%
    setnames(c("node_index", "node_id", "node_type", "node_source"), c("y_index", "y_id", "y_type", "y_source")) %>%
    setcolorder(new_colorder)
  
  # return supercluster markers
  return(supercluster_markers)
}

# get neuron supercluster to gene edges
neuron_cell_type_to_marker_gene_edges = get_cell_type_to_marker_gene_edges("neuron")
non_neuron_cell_type_to_marker_gene_edges = get_cell_type_to_marker_gene_edges("non_neuron")
cell_type_to_marker_gene_edges = rbind(neuron_cell_type_to_marker_gene_edges,  non_neuron_cell_type_to_marker_gene_edges)

# print message
message("Unique Cell Types: ", length(unique(cell_type_to_marker_gene_edges$x_id)))
message("Cell Type --> Marker Gene Edges: ", nrow(cell_type_to_marker_gene_edges))
```

Construct cell cluster to marker gene edges. Select top 50 genes.

```{r cluster-gene}
# dataset can be "neuron" or "non_neuron"
get_cell_cluster_to_marker_gene_edges = function(dataset = "neuron") {
  
  # read cluster markers
  if(dataset == "neuron") {
    cluster_markers = imap(neuron_cluster_marker_files, ~fread(here(cell_type_dir, "neuron_markers", .x)))
  } else {
    cluster_markers = imap(non_neuron_cluster_marker_files, ~fread(here(cell_type_dir, "non_neuron_markers", .x)))
  }
  
  # get cluster markers using function defined above
  # check with length(unique(non_neurons_cells$cluster_id))
  cluster_markers = map_dfr(cluster_markers, ~get_top_marker_genes(.x, n_genes = 200))
  
  # process edges
  cluster_markers = copy(cluster_markers) %>%
    .[, relation := "cell_cluster_marker_gene"] %>%
    .[, display_relation := "expression present"] %>%
    .[, x_id := gsub("_names", "", x_id, fixed = T)] %>%
    .[, x_type := "cell_cluster"] %>%
    .[, x_name := paste("cluster", x_id)] %>%
    .[, x_source := "Siletti et al."] %>%
    merge(neuroKG_nodes[node_type == "gene/protein"], by.x = "y_name", by.y = "node_name", all.x = F, all.y = F, sort = F) %>%
    setnames(c("node_index", "node_id", "node_type", "node_source"), c("y_index", "y_id", "y_type", "y_source")) %>%
    setcolorder(new_colorder)
  
  # return cluster markers
  return(cluster_markers)
}

# get neuron supercluster to gene edges
neuron_cell_cluster_to_marker_gene_edges = get_cell_cluster_to_marker_gene_edges("neuron")
non_neuron_cell_cluster_to_marker_gene_edges = get_cell_cluster_to_marker_gene_edges("non_neuron")
cell_cluster_to_marker_gene_edges = rbind(neuron_cell_cluster_to_marker_gene_edges,  non_neuron_cell_cluster_to_marker_gene_edges)

# print message
message("Unique Cell Clusters: ", length(unique(cell_cluster_to_marker_gene_edges$x_id)))
message("Cell Cluster --> Marker Gene Edges: ", nrow(cell_cluster_to_marker_gene_edges))
```

Construct cell subcluster to marker gene edges. Select top 50 genes.

```{r subcluster-gene}
# dataset can be "neuron" or "non_neuron"
get_cell_subcluster_to_marker_gene_edges = function(dataset = "neuron") {
  
  # read cluster markers
  if(dataset == "neuron") {
    subcluster_markers = imap(neuron_subcluster_marker_files, ~fread(here(cell_type_dir, "neuron_markers", .x)))
  } else {
    subcluster_markers = imap(non_neuron_subcluster_marker_files, ~fread(here(cell_type_dir, "non_neuron_markers", .x)))
  }
  
  # get cluster markers using function defined above
  # check with length(unique(non_neurons_cells$cluster_id))
  subcluster_markers = map_dfr(subcluster_markers, ~get_top_marker_genes(.x, n_genes = 50))
  
  # process edges
  subcluster_markers = copy(subcluster_markers) %>%
    .[, relation := "cell_subcluster_marker_gene"] %>%
    .[, display_relation := "expression present"] %>%
    .[, x_id := gsub("_names", "", x_id, fixed = T)] %>%
    .[, x_type := "cell_subcluster"] %>%
    .[, x_name := paste("subcluster", x_id)] %>%
    .[, x_source := "Siletti et al."] %>%
    merge(neuroKG_nodes[node_type == "gene/protein"], by.x = "y_name", by.y = "node_name", all.x = F, all.y = F, sort = F) %>%
    setnames(c("node_index", "node_id", "node_type", "node_source"), c("y_index", "y_id", "y_type", "y_source")) %>%
    setcolorder(new_colorder)
  
  # return cluster markers
  return(subcluster_markers)
}

# get neuron supercluster to gene edges
neuron_cell_subcluster_to_marker_gene_edges = get_cell_subcluster_to_marker_gene_edges("neuron")
non_neuron_cell_subcluster_to_marker_gene_edges = get_cell_subcluster_to_marker_gene_edges("non_neuron")
cell_subcluster_to_marker_gene_edges = rbind(neuron_cell_subcluster_to_marker_gene_edges,  non_neuron_cell_subcluster_to_marker_gene_edges)

# print message
message("Unique Cell Subclusters: ", length(unique(cell_subcluster_to_marker_gene_edges$x_id)))
message("Cell Subcluster --> Marker Gene Edges: ", nrow(cell_subcluster_to_marker_gene_edges))
```


# Construct Nodes

Construct nodes and populate `node_index` values; then, add node indices back to edges by merging as needed. There are 14 brain regions, 112 brain structures, 31 cell types, 450 cell clusters, and 3244 cell subclusters, so there should be $14 + 112 + 31 + 450 + 3244 = 3851$ new nodes.

```{r construct-nodes}
# get data.table of all new nodes created
all_new_nodes = rbind(
  # brain region to brain structure
  region_to_structure_edges[, .(x_id, x_type, x_name, x_source)],
  region_to_structure_edges[, .(y_id, y_type, y_name, y_source)],
  # brain structure to cell cluster
  structure_to_cell_cluster_edges[, .(x_id, x_type, x_name, x_source)],
  structure_to_cell_cluster_edges[, .(y_id, y_type, y_name, y_source)],
  # cell type to cell cluster
  cell_type_to_cell_cluster_edges[, .(x_id, x_type, x_name, x_source)],
  cell_type_to_cell_cluster_edges[, .(y_id, y_type, y_name, y_source)],
  # cell cluster to cell subcluster
  cell_cluster_to_cell_subcluster_edges[, .(x_id, x_type, x_name, x_source)],
  cell_cluster_to_cell_subcluster_edges[, .(y_id, y_type, y_name, y_source)],
  # cell type/cluster/subcluster to marker gene 
  cell_type_to_marker_gene_edges[, .(x_id, x_type, x_name, x_source)],
  cell_cluster_to_marker_gene_edges[, .(x_id, x_type, x_name, x_source)],
  cell_subcluster_to_marker_gene_edges[, .(x_id, x_type, x_name, x_source)],
  use.names = FALSE
) %>%
  unique() %>%
  setnames(c("x_id", "x_type", "x_name", "x_source"), c("node_id", "node_type", "node_name", "node_source"))
  
# print new nodes created
message("New Nodes Created: ", nrow(all_new_nodes))

# get max index and number of nodes
max_index = max(neuroKG_nodes$node_index) + 1
n_nodes = nrow(all_new_nodes)

# add node index according to NeuroKG syntax
all_new_nodes = all_new_nodes %>%
  .[, node_index := max_index:(max_index + n_nodes - 1)] %>%
  setcolorder("node_index")

# print head
# head(all_new_nodes)
```

# Add Indices

Add node indices to the edges to comply with the NeuroKG syntax.

```{r add-indices}
# function to add node indices given the x node type and y node type
add_node_indices = function(edge_data_frame, x_node_type = NULL, y_node_type = NULL) {
  
  if(is.null(x_node_type) & is.null(y_node_type)) stop("Indices may be added to only x nodes, only y nodes, or both, but not none.")
  
  # make copy
  idx_edge_data_frame = copy(edge_data_frame) %>%
    .[, x_id := as.character(x_id)] %>%
    .[, y_id := as.character(y_id)]
  
  # if needed, add x_index
  if(!is.null(x_node_type)) {
    idx_edge_data_frame = idx_edge_data_frame %>%
      merge(all_new_nodes[node_type == x_node_type, .(node_id, node_index)], by.x = "x_id", by.y = "node_id", all.x = T, all.y = F, sort = F) %>%
      setnames("node_index", "x_index")
  }
  
  # if needed, add y_index
  if(!is.null(y_node_type)) {
    idx_edge_data_frame = idx_edge_data_frame %>%
      merge(all_new_nodes[node_type == y_node_type, .(node_id, node_index)], by.x = "y_id", by.y = "node_id", all.x = T, all.y = F, sort = F) %>%
      setnames("node_index", "y_index")
  }
  
  return(setcolorder(idx_edge_data_frame, colnames(neuroKG_edges)))
  
}

# brain region to brain structure
idx_region_to_structure_edges = add_node_indices(region_to_structure_edges, "brain_region", "brain_structure")

# brain structure to cell cluster
idx_structure_to_cell_cluster_edges = add_node_indices(structure_to_cell_cluster_edges, "brain_structure", "cell_cluster")

# cell type to cell cluster
idx_cell_type_to_cell_cluster_edges = add_node_indices(cell_type_to_cell_cluster_edges, "cell_type", "cell_cluster")

# cell cluster to cell subcluster
idx_cell_cluster_to_cell_subcluster_edges = add_node_indices(cell_cluster_to_cell_subcluster_edges, "cell_cluster", "cell_subcluster")

# cell type/cluster/subcluster to marker gene
# note that y_index for gene/protein nodes has already been added
idx_cell_type_to_marker_gene_edges = add_node_indices(cell_type_to_marker_gene_edges, "cell_type")
idx_cell_cluster_to_marker_gene_edges = add_node_indices(cell_cluster_to_marker_gene_edges, "cell_cluster")
idx_cell_subcluster_to_marker_gene_edges = add_node_indices(cell_subcluster_to_marker_gene_edges, "cell_subcluster")

# bind all edges
all_new_edges = rbind(
  idx_region_to_structure_edges,
  idx_structure_to_cell_cluster_edges,
  idx_cell_type_to_cell_cluster_edges,
  idx_cell_cluster_to_cell_subcluster_edges,
  idx_cell_type_to_marker_gene_edges,
  idx_cell_cluster_to_marker_gene_edges,
  idx_cell_subcluster_to_marker_gene_edges
)

# print new nodes created
message("New Edges Created: ", nrow(all_new_edges))
```

# Construct Reverse Edges

Since graph is undirected, construct and merge reverse edges.

```{r construct-reverse}
# copy edges, where ^ indictes start of character
reverse_edges = copy(all_new_edges)
names(reverse_edges) = gsub("^x_", "temp_", names(reverse_edges))
names(reverse_edges) = gsub("^y_", "x_", names(reverse_edges))
names(reverse_edges) = gsub("^temp_", "y_", names(reverse_edges))
setcolorder(reverse_edges, colnames(all_new_edges))

# merge edges with original
all_new_edges_undirected = rbind(all_new_edges, reverse_edges, use.names = T)
```

# Add to NeuroKG

Add the newly-constructed nodes and edges to NeuroKG.

```{r save-neurokg}
# add nodes to neuroKG
neuroKG_nodes = rbind(neuroKG_nodes, all_new_nodes)

# add edges to neuroKG
neuroKG_edges = rbind(neuroKG_edges, all_new_edges_undirected)

# print number of nodes and edges
message("NeuroKG Nodes:\t", nrow(neuroKG_nodes))
message("NeuroKG Edges:\t", nrow(neuroKG_edges))

# save to file
saveRDS(neuroKG_nodes, here(neuroKG_dir, "2_enrich_KG", "neuroKG_nodes.RDS"))
saveRDS(neuroKG_edges, here(neuroKG_dir, "2_enrich_KG", "neuroKG_edges.RDS"))

# write to CSV
fwrite(neuroKG_nodes, here(neuroKG_dir, "2_enrich_KG", "neuroKG_nodes.csv"))
fwrite(neuroKG_edges, here(neuroKG_dir, "2_enrich_KG", "neuroKG_edges.csv"))
```

# Create Cell Signature Figure

Create enrichment plots that show that cell-type marker genes are enriched in the list of all genes.

```{r signature-enrichment}
# read supercluster markers
non_neuron_supercluster_markers = here(cell_type_dir, "non_neuron_markers", "supercluster_markers.csv") %>%
  fread()

# get astrocyte markers
astrocyte_markers = non_neuron_supercluster_markers %>%
  .[, .SD, .SDcols = grep("Astrocyte", colnames(.), value = T)] %>%
  .[, Ranking := 1:nrow(.)]

# read astrocyte gene sets
astrocyte_gene_sets = here("Data", "cell_type", "gene_sets", "astrocyte_gene_sets.csv") %>%
  fread()

# get ADRA gene set and scores
ADRA_genes = astrocyte_gene_sets$`Serrano-Pozo_ADRA` %>% .[!(. == "")]
ADRA_gene_scores = astrocyte_markers[Astrocyte_names %in% ADRA_genes]

# create barcode plot
# pdf(here(neuroKG_dir, "2_enrich_KG", "astrocyte_enrichment_plot.pdf"), width = 6, height = 8)
# limma::barcodeplot(astrocyte_markers$Astrocyte_scores, ADRA_gene_scores$Ranking)
# dev.off()

```
